First commit of guide_bot to git.
Mads Bertelsen started development of guide_bot in 2013 with a useable version ready the same year.
The project was included in both the Master thesis (2014) and PhD thesis (2017) of Mads Bertelsen.

Emmanouela Rantsiou contributed code relating to the PSI cluster.

Leland Harriger have made substatial contributions in the form of a Monochromator module and several other smaller changes in the main code, for example measurement of brilliance transfer at places other than the sample and beamport rotation on the Source_gen source.

The current version is named 1.04 and is a manual merge between the development version of Mads Bertelsen and Leland Harriger.