Update to mcstas_bot.m to avoid using the iFit data field Mean.
The meta data Mean field was removed from McStas data in newer
version of iFit, and thus the guide_bot analysis files now
calculate the mean with the ifit_mean function that is provided.
This should allow a wider range of MATLAB / iFit combinations,
though it was only tested with MATLAB 2016 / iFit 1.8.

Some cleanup of files not used by guide_bot was performed as well.